+++ src/Text/Show/Text/Compiler.hs
@@ -1,0 +1,13 @@
+{-|
+Module:      Text.Show.Text.Compiler
+Copyright:   (C) 2014-2015 Ryan Scott
+License:     BSD-style (see the file LICENSE)
+Maintainer:  Ryan Scott
+Stability:   Experimental
+Portability: GHC
+
+Imports 'Show' instances for @Compiler@ modules.
+-}
+module Text.Show.Text.Compiler () where
+
+import Text.Show.Text.Compiler.Hoopl ()