+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.2
-* Added instances for the `binary` and `utf8-string` libraries
+* Added instances for the `binary`, `hoopl`, and `utf8-string` libraries
 
 # 0.1
 * Initial commit